+++ backend/data/src/main/kotlin/io/tolgee/model/enums/Announcement.kt
@@ -5,7 +5,8 @@ import java.time.format.DateTimeFormatter
 
 @Suppress("unused")
 enum class Announcement(val until: Long) {
-  FEATURE_BATCH_OPERATIONS(parseTime("2023-09-10 00:00 UTC"));
+  FEATURE_BATCH_OPERATIONS(parseTime("2023-09-10 00:00 UTC")),
+  FEATURE_MT_FORMALITY(parseTime("2023-10-20 00:00 UTC"));
 
   companion object {
     val last: Announcement